Extract method for populating the criteria state
This is done to increase readability and make method/file less complex
